+++ src/Contacts/src/SchmittTriggerDetector.cpp
@@ -7,7 +7,6 @@
 
 #include <BipedalLocomotion/ContactDetectors/SchmittTriggerDetector.h>
 
-using namespace BipedalLocomotion::Estimators;
 using namespace BipedalLocomotion::ParametersHandler;
 using namespace BipedalLocomotion::Contacts;